insert-data/requirements.txt
Faker==37.12.0
mysql-connector-python==9.5.0
tzdata==2025.2
